+++ compiler/src/main/kotlin/software/amazon/lastmile/kotlin/inject/anvil/processor/ContributesBindingProcessor.kt
@@ -38,7 +38,7 @@ import kotlin.reflect.KClass
  *
  * @Inject
  * @SingleIn(AppScope::class)
- * @ContributesBinding
+ * @ContributesBinding(AppScope::class)
  * class RealAuthenticator : Authenticator
  * ```
  * Will generate: