+++ compiler/src/test/kotlin/software/amazon/test/Singleton.kt
@@ -1,7 +1,0 @@
-package software.amazon.test
-
-import me.tatarka.inject.annotations.Scope
-
-@Scope
-@Target(AnnotationTarget.CLASS, AnnotationTarget.FUNCTION, AnnotationTarget.PROPERTY_GETTER)
-annotation class Singleton